# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Lead Service",
    "description": "Manages lead buffering, deduplication, cursor pagination, and enrichment for campaign outreach.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3006"
    }
  ],
  "security": [
    {
      "apiKey": []
    }
  ],
  "components": {
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "service": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "service"
        ]
      },
      "BufferPushResponse": {
        "type": "object",
        "properties": {
          "buffered": {
            "type": "number"
          },
          "skippedAlreadyServed": {
            "type": "number"
          }
        },
        "required": [
          "buffered",
          "skippedAlreadyServed"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "BufferPushRequest": {
        "type": "object",
        "properties": {
          "campaignId": {
            "type": "string"
          },
          "brandId": {
            "type": "string"
          },
          "parentRunId": {
            "type": "string"
          },
          "clerkUserId": {
            "type": "string"
          },
          "leads": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "externalId": {
                  "type": "string",
                  "nullable": true
                },
                "data": {
                  "nullable": true
                }
              },
              "required": [
                "email"
              ]
            }
          }
        },
        "required": [
          "campaignId",
          "brandId",
          "leads"
        ]
      },
      "BufferNextResponse": {
        "type": "object",
        "properties": {
          "found": {
            "type": "boolean"
          },
          "lead": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string"
              },
              "externalId": {
                "type": "string",
                "nullable": true
              },
              "data": {
                "nullable": true
              },
              "brandId": {
                "type": "string"
              },
              "clerkOrgId": {
                "type": "string",
                "nullable": true
              },
              "clerkUserId": {
                "type": "string",
                "nullable": true
              }
            },
            "required": [
              "email",
              "externalId",
              "brandId",
              "clerkOrgId",
              "clerkUserId"
            ]
          }
        },
        "required": [
          "found"
        ]
      },
      "BufferNextRequest": {
        "type": "object",
        "properties": {
          "campaignId": {
            "type": "string"
          },
          "brandId": {
            "type": "string"
          },
          "parentRunId": {
            "type": "string"
          },
          "searchParams": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            }
          },
          "clerkUserId": {
            "type": "string"
          }
        },
        "required": [
          "campaignId",
          "brandId"
        ]
      },
      "CursorGetResponse": {
        "type": "object",
        "properties": {
          "state": {
            "nullable": true
          }
        }
      },
      "CursorSetResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
      "CursorSetRequest": {
        "type": "object",
        "properties": {
          "state": {
            "nullable": true
          }
        }
      },
      "LeadsResponse": {
        "type": "object",
        "properties": {
          "leads": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": {
                "nullable": true
              }
            }
          }
        },
        "required": [
          "leads"
        ]
      },
      "StatsResponse": {
        "type": "object",
        "properties": {
          "served": {
            "type": "number"
          },
          "buffered": {
            "type": "number"
          },
          "skipped": {
            "type": "number"
          },
          "apollo": {
            "type": "object",
            "properties": {
              "enrichedLeadsCount": {
                "type": "number"
              },
              "searchCount": {
                "type": "number"
              },
              "fetchedPeopleCount": {
                "type": "number"
              },
              "totalMatchingPeople": {
                "type": "number"
              }
            },
            "required": [
              "enrichedLeadsCount",
              "searchCount",
              "fetchedPeopleCount",
              "totalMatchingPeople"
            ]
          }
        },
        "required": [
          "served",
          "buffered",
          "skipped",
          "apollo"
        ]
      },
      "StatsPostRequest": {
        "type": "object",
        "properties": {
          "runIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "appId": {
            "type": "string"
          },
          "brandId": {
            "type": "string"
          },
          "campaignId": {
            "type": "string"
          },
          "clerkOrgId": {
            "type": "string"
          }
        }
      }
    },
    "parameters": {},
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for authenticating requests. Must match the LEAD_SERVICE_API_KEY env var on the server."
      }
    }
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/buffer/push": {
      "post": {
        "summary": "Push leads into the buffer",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "API key for authenticating requests"
          },
          {
            "in": "header",
            "name": "x-app-id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Identifies the calling application, e.g. mcpfactory"
          },
          {
            "in": "header",
            "name": "x-org-id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "External organization ID, e.g. Clerk org ID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BufferPushRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Leads buffered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BufferPushResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/buffer/next": {
      "post": {
        "summary": "Pull the next lead from the buffer",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "API key for authenticating requests"
          },
          {
            "in": "header",
            "name": "x-app-id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Identifies the calling application, e.g. mcpfactory"
          },
          {
            "in": "header",
            "name": "x-org-id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "External organization ID, e.g. Clerk org ID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BufferNextRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Next lead from buffer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BufferNextResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/cursor/{namespace}": {
      "get": {
        "summary": "Get cursor state for a namespace",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "API key for authenticating requests"
          },
          {
            "in": "header",
            "name": "x-app-id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Identifies the calling application, e.g. mcpfactory"
          },
          {
            "in": "header",
            "name": "x-org-id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "External organization ID, e.g. Clerk org ID"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "namespace",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Cursor state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorGetResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "summary": "Set cursor state for a namespace",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "API key for authenticating requests"
          },
          {
            "in": "header",
            "name": "x-app-id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Identifies the calling application, e.g. mcpfactory"
          },
          {
            "in": "header",
            "name": "x-org-id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "External organization ID, e.g. Clerk org ID"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "namespace",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CursorSetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cursor updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorSetResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/leads": {
      "get": {
        "summary": "List served leads with enrichment data",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "API key for authenticating requests"
          },
          {
            "in": "header",
            "name": "x-app-id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Identifies the calling application, e.g. mcpfactory"
          },
          {
            "in": "header",
            "name": "x-org-id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "External organization ID, e.g. Clerk org ID"
          },
          {
            "in": "query",
            "name": "brandId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "clerkOrgId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "clerkUserId",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of served leads",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/stats": {
      "get": {
        "summary": "Get lead stats by status",
        "description": "Returns counts of leads by status: served (delivered with verified email), buffered (awaiting enrichment), and skipped (no email found).",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "API key for authenticating requests"
          },
          {
            "in": "header",
            "name": "x-app-id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Identifies the calling application, e.g. mcpfactory"
          },
          {
            "in": "header",
            "name": "x-org-id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "External organization ID, e.g. Clerk org ID"
          },
          {
            "in": "query",
            "name": "brandId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "campaignId",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lead stats by status with Apollo metrics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Get lead stats by status (internal)",
        "description": "Service-to-service endpoint. Returns counts of leads by status.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatsPostRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Lead stats by status with Apollo metrics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/openapi.json": {
      "get": {
        "summary": "Get OpenAPI specification",
        "responses": {
          "200": {
            "description": "OpenAPI JSON document"
          },
          "404": {
            "description": "Spec not generated"
          }
        }
      }
    }
  }
}